refactor: split pipeline.ts into modular stage files
Split the monolithic pipeline.ts into pipeline/ directory with one file
per stage, preparing for future expansion (e.g. DB write-ahead buffer):

  pipeline/
  ├── index.ts          — composition + architecture docs
  ├── persist.ts        — order 10: DB persistence
  ├── ring-buffer.ts    — order 20: in-memory buffer
  ├── auto-title.ts     — order 30: title extraction
  └── failure-detect.ts — order 40: logical failure detection

Each stage exports a register function that receives the event bus `on`
helper and EngineContext. The index composes all stages and owns the
pipeline-level documentation.

No behavioral changes — engine.ts import resolves automatically via
directory index.
